+++ gst-plugin-http/src/httpsrc.rs
@@ -13,6 +13,8 @@ use std::io::Read;
 use std::u64;
 use url::Url;
 
+use gst_plugin::error::*;
+
 use gst_plugin_simple::error::*;
 use gst_plugin_simple::source::*;
 use gst_plugin_simple::UriValidator;
