+++ gst-plugin-simple/src/sink.rs
@@ -20,6 +20,7 @@ use gobject_subclass::object::*;
 use gst_plugin::base_sink::*;
 use gst_plugin::element::*;
 use gst_plugin::uri_handler::*;
+use gst_plugin::error::*;
 
 pub use gst_plugin::base_sink::BaseSink;
